letslearnjava.quora.com/Thread-Pool-Concept-in-Java
Thread Pool i.e. for the first time your thread will be generated as
fresh & when it will be done with its task and it will be made to store
into the pool and whenever it will be required in future directly fetch
live thread from the pool (if available), assign a task & again this
cycle continues. You don't need to start a thread after initial start
you made it for first time.